+++ services/chaindb/postgresql/epochsummaries.go
@@ -1,4 +1,4 @@
-// Copyright © 2021 Weald Technology Limited.
+// Copyright © 2021 - 2023 Weald Technology Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -15,7 +15,11 @@ package postgresql
 
 import (
 	"context"
+	"fmt"
+	"sort"
+	"strings"
 
+	"github.com/pkg/errors"
 	"github.com/wealdtech/chaind/services/chaindb"
 	"go.opentelemetry.io/otel"
 )
@@ -97,4 +101,133 @@ func (s *Service) SetEpochSummary(ctx context.Context, summary *chaindb.EpochSum
 	)
 
 	return err
+}
+
+// EpochSummaries provides summaries according to the filter.
+func (s *Service) EpochSummaries(ctx context.Context, filter *chaindb.EpochSummaryFilter) ([]*chaindb.EpochSummary, error) {
+	ctx, span := otel.Tracer("wealdtech.chaind.services.chaindb.postgresql").Start(ctx, "EpochSummaries")
+	defer span.End()
+
+	tx := s.tx(ctx)
+	if tx == nil {
+		ctx, err := s.BeginROTx(ctx)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to begin transaction")
+		}
+		defer s.CommitROTx(ctx)
+		tx = s.tx(ctx)
+	}
+
+	// Build the query.
+	queryBuilder := strings.Builder{}
+	queryVals := make([]interface{}, 0)
+
+	queryBuilder.WriteString(`
+SELECT f_epoch
+      ,f_activation_queue_length
+      ,f_activating_validators
+      ,f_active_validators
+      ,f_active_real_balance
+      ,f_active_balance
+      ,f_attesting_validators
+      ,f_attesting_balance
+      ,f_target_correct_validators
+      ,f_target_correct_balance
+      ,f_head_correct_validators
+      ,f_head_correct_balance
+      ,f_attestations_for_epoch
+      ,f_attestations_in_epoch
+      ,f_duplicate_attestations_for_epoch
+      ,f_proposer_slashings
+      ,f_attester_slashings
+      ,f_deposits
+      ,f_exiting_validators
+      ,f_canonical_blocks
+FROM t_epoch_summaries`)
+
+	wherestr := "WHERE"
+
+	if filter.From != nil {
+		queryVals = append(queryVals, *filter.From)
+		queryBuilder.WriteString(fmt.Sprintf(`
+%s f_epoch >= $%d`, wherestr, len(queryVals)))
+		wherestr = "  AND"
+	}
+
+	if filter.To != nil {
+		queryVals = append(queryVals, *filter.To)
+		queryBuilder.WriteString(fmt.Sprintf(`
+%s f_epoch <= $%d`, wherestr, len(queryVals)))
+	}
+
+	switch filter.Order {
+	case chaindb.OrderEarliest:
+		queryBuilder.WriteString(`
+ORDER BY f_epoch`)
+	case chaindb.OrderLatest:
+		queryBuilder.WriteString(`
+ORDER BY f_epoch DESC`)
+	default:
+		return nil, errors.New("no order specified")
+	}
+
+	if filter.Limit > 0 {
+		queryVals = append(queryVals, filter.Limit)
+		queryBuilder.WriteString(fmt.Sprintf(`
+LIMIT $%d`, len(queryVals)))
+	}
+
+	if e := log.Trace(); e.Enabled() {
+		params := make([]string, len(queryVals))
+		for i := range queryVals {
+			params[i] = fmt.Sprintf("%v", queryVals[i])
+		}
+		e.Str("query", strings.ReplaceAll(queryBuilder.String(), "\n", " ")).Strs("params", params).Msg("SQL query")
+	}
+
+	rows, err := tx.Query(ctx,
+		queryBuilder.String(),
+		queryVals...,
+	)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	summaries := make([]*chaindb.EpochSummary, 0)
+	for rows.Next() {
+		summary := &chaindb.EpochSummary{}
+		err := rows.Scan(
+			&summary.Epoch,
+			&summary.ActivationQueueLength,
+			&summary.ActivatingValidators,
+			&summary.ActiveValidators,
+			&summary.ActiveRealBalance,
+			&summary.ActiveBalance,
+			&summary.AttestingValidators,
+			&summary.AttestingBalance,
+			&summary.TargetCorrectValidators,
+			&summary.TargetCorrectBalance,
+			&summary.HeadCorrectValidators,
+			&summary.HeadCorrectBalance,
+			&summary.AttestationsForEpoch,
+			&summary.AttestationsInEpoch,
+			&summary.DuplicateAttestationsForEpoch,
+			&summary.ProposerSlashings,
+			&summary.AttesterSlashings,
+			&summary.Deposits,
+			&summary.ExitingValidators,
+			&summary.CanonicalBlocks,
+		)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to scan row")
+		}
+		summaries = append(summaries, summary)
+	}
+
+	// Always return order of epoch.
+	sort.Slice(summaries, func(i int, j int) bool {
+		return summaries[i].Epoch < summaries[j].Epoch
+	})
+	return summaries, nil
 }